+++ game_units/src/player.h
@@ -1,19 +1,21 @@
 #ifndef LAST_SAVIORS_PLAYER_H
 #define LAST_SAVIORS_PLAYER_H
 
+#include <cstddef>
+
 enum TowerType { AreaDamageTower, PointDamageTower };
 
 class Player {
    private:
-    const int INITIAL_COINS = 200;
-    int coins;
+    const size_t INITIAL_COINS = 200;
+    size_t coins;
 
    public:
     Player();
     void buildTower(TowerType type);
     void takePuzzle(int complexity);
-    void increaseCoins(int value);
-    void reduceCoins(int value);
+    void increaseCoins(size_t value);
+    void reduceCoins(size_t value);
     void buildShaft();
 };
